consuela/config/requirements.txt
pandas>=1.3.0
pyyaml>=6.0
pyarrow>=5.0.0

# Data Downloading Dependencies
yfinance>=0.2.0
fredapi>=0.5.0
newsapi-python>=0.2.0
pytrends>=4.9.0
textblob>=0.17.0
requests>=2.28.0

# Ramanujan ML Framework Dependencies
numpy>=1.21.0
scikit-learn>=1.0.0
scipy>=1.7.0
joblib>=1.1.0

# Gradient Boosting Models
xgboost>=1.6.0
lightgbm>=3.3.0
catboost>=1.0.0

# Deep Learning
tensorflow>=2.8.0

# Time Series and Financial Models
arch>=5.3.0

# Sequence Models
hmmlearn>=0.2.7

# Visualization (optional)
matplotlib>=3.5.0
seaborn>=0.11.0

# Hyperparameter Optimization (optional)
optuna>=3.0.0

# Data Download Dependencies
yfinance>=0.2.0
fredapi>=0.5.0
requests>=2.28.0